--- src/main/java/Command.java
@@ -1,7 +1,0 @@
-public abstract class Command {
-    abstract void executeCommand(TaskList tasks, Ui ui, Storage storage);
-    
-    public boolean isExit() {
-        return false;
-    }
-}
