+++ src/main/java/duke/command/ByeCommand.java
@@ -1,3 +1,9 @@
+package duke.command;
+
+import duke.storage.Storage;
+import duke.task.TaskList;
+import duke.ui.Ui;
+
 public class ByeCommand extends Command {
     @Override
     public void executeCommand(TaskList tasks, Ui ui, Storage storage) {
