+++ src/main/java/duke/command/DeadlineCommand.java
@@ -1,3 +1,10 @@
+package duke.command;
+
+import duke.storage.Storage;
+import duke.task.Deadline;
+import duke.task.Task;
+import duke.task.TaskList;
+import duke.ui.Ui;
 import java.time.LocalDate;
 
 public class DeadlineCommand extends Command {
