+++ src/main/java/duke/command/EventCommand.java
@@ -1,3 +1,10 @@
+package duke.command;
+
+import duke.storage.Storage;
+import duke.task.Event;
+import duke.task.Task;
+import duke.task.TaskList;
+import duke.ui.Ui;
 import java.time.LocalDate;
 
 public class EventCommand extends Command {
