+++ src/main/java/duke/task/Deadline.java
@@ -1,22 +1,24 @@
+package duke.task;
+
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
 
-public class Event extends Task {
+public class Deadline extends Task {
     private LocalDate time;
 
-    public Event(String description, boolean done, LocalDate time) {
+    public Deadline(String description, boolean done, LocalDate time) {
         super(description, done);
         this.time = time;
     }
 
     @Override
     public String toString() {
-        return "[E]" + super.toString() + " (at: " +
+        return "[D]" + super.toString() + " (by: " +
                 this.time.format(DateTimeFormatter.ofPattern("MMM d yyyy")) + ")";
     }
 
     @Override
     public String saveToHardDisk() {
-        return "E" + super.saveToHardDisk() + " | " + this.time;
+        return "D" + super.saveToHardDisk() + " | " + this.time;
     }
 }
