+++ src/main/java/duke/task/Event.java
@@ -1,22 +1,24 @@
+package duke.task;
+
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
 
-public class Deadline extends Task {
+public class Event extends Task {
     private LocalDate time;
 
-    public Deadline(String description, boolean done, LocalDate time) {
+    public Event(String description, boolean done, LocalDate time) {
         super(description, done);
         this.time = time;
     }
 
     @Override
     public String toString() {
-        return "[D]" + super.toString() + " (by: " +
+        return "[E]" + super.toString() + " (at: " +
                 this.time.format(DateTimeFormatter.ofPattern("MMM d yyyy")) + ")";
     }
 
     @Override
     public String saveToHardDisk() {
-        return "D" + super.saveToHardDisk() + " | " + this.time;
+        return "E" + super.saveToHardDisk() + " | " + this.time;
     }
 }
