+++ src/main/java/duke/task/ToDo.java
@@ -1,3 +1,5 @@
+package duke.task;
+
 public class ToDo extends Task {
     public ToDo(String description, boolean done) {
         super(description, done);
